LPD-42702 We need to perform the delete inside the try because we are checking the deletion of accounts

+++ modules/apps/account/account-test/src/testIntegration/java/com/liferay/account/internal/model/listener/test/CompanyModelListenerTest.java
@@ -56,6 +56,10 @@ public class CompanyModelListenerTest {
 				accountEntry.getAccountEntryId(), RandomTestUtil.randomString(),
 				null, null);
 
+			_companyLocalService.deleteCompany(company);
+
+			company = null;
+
 			Assert.assertNull(
 				_accountEntryLocalService.fetchAccountEntry(
 					accountEntry.getAccountEntryId()));
@@ -65,7 +69,9 @@ public class CompanyModelListenerTest {
 					accountRole.getAccountRoleId()));
 		}
 		finally {
-			_companyLocalService.deleteCompany(company);
+			if (company != null) {
+				_companyLocalService.deleteCompany(company);
+			}
 		}
 	}
